chat geo added
