+++ src/app/shared/geo-chart/geo-chart.component.html
@@ -1,0 +1,1 @@
+<div #chartContainer></div>
