+++ src/app/shared/shared.module.ts
@@ -1,15 +1,18 @@
 import { NgModule } from '@angular/core';
 import { ChartComponent } from './chart/chart.component';
+import { GeoChartComponent } from './geo-chart/geo-chart.component';
 
 
 
 @NgModule({
   declarations: [],
   imports: [
-    ChartComponent
+    ChartComponent,
+    GeoChartComponent
   ],
   exports: [
-    ChartComponent
+    ChartComponent,
+    GeoChartComponent
   ]
 })
 export class SharedModule { }
